Call Connect -- Update call status
This endpoint is to be run when a desktop app user clicks on a 'Connect' button. It's purpose
is to track call status and notify other clients of the call.

Voice connection occurs separately, client side in thought of improved performance.
Concerns exist over leaking auth tokens across Desktop App, and a lag between connected
calls, and GUI in the Desktop App. An alternative is to return OpenTok tokens in this endpoint